`verdi quicksetup`: Fix error incorrect role when creating database (#5828)
In some postgresql installations, the following fails:

    template1=> create user test;
    CREATE ROLE
    template1=> create database test owner test;
    ERROR:  must be member of role "test"

This is fixed by:

    template1=> grant test to azureuser;
    GRANT ROLE
    template1=> create database test owner test;
    CREATE DATABASE

Furthermore, granting privileges to the owner is superfluous according
to the postgresql docs, so we drop it.

The investigation of this issue also surfaces a bug in the error message
for "manual database creation", and that the `--su-db-name` was not
passed on from the commandline. This is fixed in passing.