DevOps: Fix the `test-install` workflow (#5873)
Recently the `test-install` workflow started failing because `mamba` can
no longer be installed. The replacement of `conda` with `mamba` was
introduced in 6da13333222af01c15b676ed79c98855442923c3 because `conda`
was running out of memory.

The `s-weigand/setup-conda` action is replaced with the more up-to-date
`conda-incubator/setup-miniconda` action. Instead of installing `mamba`
we install `conda` to work-around the current problem. To prevent the
out-of-memory error, we still use the mamba solvers through

    conda install -n base conda-libmamba-solver
    conda config --set solver libmamba