`InstalledCode`: Allow relative path for `filepath_executable`
The `InstalledCode` required the `filepath_executable` to be an absolute
path. The original reasoning behind this design choice was that this
improves the provenance of data produced with this code slightly as it
would be more difficult to accidentally change the version of the code
that was associated with the executable. For relative paths, the actual
linked code could be easily changed through, for example, environment
variables.

However, this restriction also has downsides. On many compute clusters
the preferred approach is to load specific modules that provide the
software of interest in which relative executables are provided. The
absolute paths can change depending on microarchitecture of the system
or other system variables. Requiring the user to provide the absolute
path to the executable defeats the purpose of this system.

Therefore the requirement of the executable path having to be absolute
is dropped.