`IcsdDbImporter`: Add `is_theoretical` tag to queried entries (#5868)
The extra `is_theoretical: True/False` for ICSD entries queried from the
MySQL database is added indicating whether the structure is theoretical
or has been confirmed experimentally. A structure can be considered
theoretical when the remark `THE` or `ZTHE` is present for the
corresponding row in the `icsd_remarks` table.