`verdi quicksetup`: Further reduce required user interaction (#5768)
When `verdi quicksetup` detects that a database user or database already
exists, it prompts the user whether to use those and, if not, to pick a
new name. Novice users likely do not know what exactly this choice
entails; furthermore, choosing to reuse the database user could actually
fail, namely when the correct password for this user was not known in
the `config.json`.

Furthermore, if the user decides not to reuse the database or database
user, there is no point in letting them choose the new name. It is much
easier to autogenerate it (if the user really wants to choose the name,
they can pass it in via the `--db-username` and `--db-name` options to
`verdi quicksetup`).

Here, this user interaction is removed from `verdi quicksetup` and new
names for the database and database user are simply generated as needed.