`verdi node delete`: Add verbose list of pks to be deleted (#5878)
The `verdi node delete` command provided no way to a user to introspect
which pks would be deleted, just the total number of nodes. A log
message on the INFO level is added with the full list of pks to be
deleted. A user can use the `-v info` option to print it.