Process functions: Replace `getfullargspec` with `signature`
The `inspect.getfullargspec` method is used to analyze the signature of
the function wrapped by the process function decorator. This method is
outdated and really just kept around for backwards-compatibility.
Internally it is using the `inspect.signature` method which has a more
modern interface, so here we change the code to use that method instead.
This change also prepares the next move that will allow to parse any
type annotations for signature parameters to dynamically determine the
`valid_type` attribute of the dynamically determined input ports.