Process functions: Infer argument `valid_type` from type hints
An advantage of process functions over calculation jobs and work chains
are that they are light weight and very easy to start using. But there
are also disadvantages; due to the process specification being inferred
from the function signature, the user cannot benefit from all the
features of the process specification. For example, it is not possible
to explicitly define a help string or valid type for the arguments of
the process function.

With PEP 484 type hints were introduced in Python 3.5 that made it
possible to indicate the expected types for function arguments. Here we
make use of this functionality to infer the valid type of function
arguments based on the provided type hint, if any.

As an example, a user can now define the following calcfunction:

    @calcfunction
    def add(a: int, b: int):
        return a + b

and when it is called with anything other than an `int` or `Int` for
either of the two arguments, a validation error is raised reporting that
an argument with an invalid type was provided.

The new functionality is fully optional and process functions without
typing will continue to work as before. If incorrect typing is provided,
a warning is logged and the typing is ignored.

The annotation of the wrapped process function is parsed using the
`inspect.get_annotation` method. This was added in Python 3.10 and so to
provide support in older versions we install the `get-annotations`
backport package. Since we have to use `eval_str=True` in the call to
get unstringized versions of the types. This will fail for the
backport implementation if the type uses union syntax of PEP 604, e.g
`str | int` instead of `typing.Union[str, int]`, even if this
functionality is enabled using `from __future__ import annotations`.