Fix the output filename handling of `verdi node graph generate`
The `graphviz.Digraph.render` method would automatically add a suffix to
the output filename when specified with `filename`, which is not what we
want. Using `outfile` instead solves the problem as it doesn't modify the
filename. However, a warning will be issued if the suffix of the
filename does not match the chosen output format, but that is
acceptable.