Generate content for `verdi code show` automatically
Some CLI commands like `verdi code show` rely on printing various node
attributes 'by hand' and so may miss certain attributes and must be
maintained 'by hand' on an ongoing basis.

This commit uses `get_cli_options` to populate the output table for the
`show` subcommand automatically, so as to ensure completeness of output
and to simplify future maintenance.