Add support for Docker images to use as `Code` for `CalcJob`s (#5841)
The `ContainerizedCode` is currently not quite compatible with Docker's
containerization technology. The reason is that the executable,
including all of its command line arguments and the file descriptor
redirections need to be part of a single quoted string, passed to the
`bash -c` command that is run inside the container.

For example, to run `pw.x` the submit line needs to look like:

    docker run -i {image_name} sh -c "pw.x -input input.in"

To enable this, a new attribute is added to the `JobTemplateCodeInfo`
dataclass. This instructs the `Scheduler` to wrap the executable and all
its arguments in quotes. The `AbstractCode` has a new attribute with the
same name that is `False` by default to keep backwards-compatibility
but which can be set to `True` to enable compatiblity with Docker.

The attribute has an associated getter and setter, which is added to the
`AbstractCode` and not the `ContainerizedCode` because the `Scheduler`
plugin will attempt to retrieve this attribute for all code types. If
the properties would be added to `ContainerizedCode`, the scheduler
plugin would except with an `AttributeError`.