Process functions: Parse docstring to set input port help attribute
The `docstring-parser` package is used to parse the docstring of a
process function if one is present. If the parsing is successful the
description of the function arguments is used to set the `help`
attribute of the corresponding `InputPort` that is added to the
dynamically generated process specification.

This is especially useful when the process function is exposed in a
wrapping workchain as now the description of the inputs is inherited
from the docstring and becomes immediately available on the process spec
of the workchain. This allows a user to inspect the inputs description
without having to go look at the source code of the function.