Docs: Clarify that process functions can be exposed in work chains
This functionality was already supported, given that process functions
automatically have a `Process` class generated from the function
signature, but it was never documented explicitly. It now also mentions
that inferred `valid_type` and `help` attributes on the input ports as
inferred from the type hints and docstrings, respectively, are
preserverd when exposing. An explicit test has been added for this.