`verdi config list`: Do not except if no profiles are defined
The command should work if no profiles exist, in which case it should
simply list the global values, however, it was excepting since it
assumed that `ctx.obj.profile` would be set. This is set by the
`ProfileParamType` which is used by the `--profile` option, but this is
not called if no profiles are defined.

To create a regression test, the `empty_config` fixture is used.
However, since this fixture works in memory, the `use_subprocess=False`
switch has to be used for `run_cli_command`. This by default would
populate the `ctx.obj.profile` object. This is normally done by the
`--profile` option, but this is not hit when using `click.CliRunner`.
This would nullify the test, as it would always pass, even without the
fix.

To solve this issue, a new argument `initialize_ctx_obj` is added to the
`run_cli_command` fixture. If set to `False`, the `ctx.obj` is not
initialized. By using this new argument, the regression test properly
failed before the test, but passes after the fix.