`AbstractCode`: Add the `with_mpi` attribute
The `with_mpi` attribute serves to indicate whether a code instance
needs to be run with the MPI run command in front of the executable.
The attribute is exposed through the `with_mpi` property getter and
setter. The attribute can be set through the `verdi code create` command
using the `--with-mpi/--no-with-mpi` flags.

The attribute is set to `None` by default. This means that the code
doesn't enforce MPI being used or not whatsoever and it is left up to
the `CalcJob` plugin and the `metadata.options.withmpi` input.

The `CalcJob.presubmit` logic is updated to take the `with_mpi` setting
of the input codes into account. The default value whether to run with
or without MPI is defined by the `metadata.options.withmpi` input. This
can then be overridden by either the plugin, code input or through the
`metadata.options.withmpi` input, but if multiple are explicitly
specified, they have to agree or a `RuntimeError` is raised.