`WorkChains`: Raise if predicate `if_/while_` does not return boolean (#5924)
Before the conditionals created by the `if_` and `while_` constructs
would enter their body as long as the predicate returned anything that
was truthy. This could lead to unexpected behavior. For example, if a
user tried to abort a workchain from the predicate of a while-loop by
returning an `ExitCode` instance, the while stepper would accept the
exit code as truthy and continue the loop.

To prevent this, the conditionals will now raise a `TypeError` whenever
the value returned by the predicate is not a boolean. Since this logic
was added in `plumpy==0.21.4` the dependency requirement is updated.