`verdi code create`: Do not prompt for `--with-mpi`
The `--with-mpi/--no-with-mpi` flag allows to set the `with_mpi`
attribute on the code. This attribute is optional and allows `None` as
a default but since the option is a boolean flag, in interactive mode,
it prompts and forces the user to choose `True` or `False` and it is not
possible to leave it unset, which should be the default behavior.

To solve this, `DynamicEntryPointCommandGroup.create_option` method,
which is used to dynamically generate the CLI options for the various
code subclasses, is updated to remove the prompt for flag options if the
default is `None`. Now, the option will no longer prompt in interactive
mode (and set `None` as the default), but the flag can still be used in
non-interactive mode.