Pre-commit: Set `yapf` option `allow_split_before_dict_value = false`
This option prevents linebreaks after keys in dictionaries if the values
exceed the maximum column width. This happens for example in
dictionaries where values are long multiline strings. If the string
exceeds the max length, it is put on the next line, which is not too
bad, but so are all other values in the dictionary, even those that are
not too long. This leads to difficult to read formatting.