CLI: Remove stale PID file deleting from status commands
Summary of behavior changes:

* `verdi daemon status` and `verdi status` no longer delete stale PID file
* `verdi daemon start/restart/stop` now automatically clean stale PID file

The `verdi status` and `verdi daemon status` commands were calling the
function `aiida.cmdline.utils.daemon.delete_stale_pid_file` which would
check the daemon PID file and delete it if it was considered stale. This
was added to prevent users having to clean up manually if the daemon got
killed abruptly, leaving behind an orphaned PID file. However, this is a
side-effect that users wouldn't expect from a command that should simply
report the status of a service.

The check and removal of a stale PID file is removed from the status
commands, and is instead moved to the `stop_daemon`, start_daemon` and
`restart_daemon` of the `DaemonClient`. These methods are called by the
`verdi daemon stop/start/restart` commands. These commands are designed
to change the state of the daemon, so if they automatically clean a
seemingly stale daemon PID, that is to expected behavior.

The `aiida.cmdline.utils.daemon.delete_stale_pid_file` function has been
removed and its functionality has been refactored and moved to the
`DaemonClient._check_pid_file` method. It raises if the PID file is
likely to be stale. This allows it to be wrapped by `_is_pid_file_stale`
which returns a boolean, and `_clean_potentially_stale_pid_file` which
replaces the old functionality of automatically deleting the PID file if
it is stale and logs a warning of having done so.

The `aiida.cmdline.utils.daemon.get_status` function has also been
removed. It was only used by `verdi daemon status` and `verdi status`
where the latter even just stripped out the first line to just get a
string to say the daemon is running. Part of the functionality is
replaced by `DaemonClient.get_status` which returns a dataclass
`DaemonStatus` that contains whether the daemon is running, the
connection errored out and a human-readable message. This is sufficient
for `verdi status`. The logic that formatted additional information
about daemon workers has been moved to `verdi daemon start` itself.

The `daemon.timeout` option default is changed from 20 to 5 seconds.
There is no reason why a functioning daemon wouldn't be able to respond
within 5 seconds, and forcing a user to wait 20 seconds in case of a
broken daemon is frustratingly long.