`DaemonClient`: Refactor to include parsing of client response
Recently, the `DaemonClient` has been updated to provide all necessary
functionality to manage the daemon through the Python API just as it is
possible through `verdi daemon`. For example, the methods `start_daemon`
and `stop_daemon` were added. However, the interface of the client is
still lacking, particularly because the various methods returned a
dictionary from which the caller would still have to parse the actual
result from string values.

The `call_client` method is updated to detect various problems and
communicate them more clearly by raising an exception. Various
subclasses of `DaemonException` are raised for certain problems:

* `DaemonNotRunningException`: If the daemon is not running
* `DaemonStalePidException`: If the daemon could not be reached and the
  PID file appears to be stale.
* `DaemonTimeoutException`: If the daemon connection timed out
* `DaemonException`: For any other (unknown) reason

The functionality that checks for a stale PID file is moved from the
`aiida.cmdline.utils.daemon` module to the `DaemonClient`. This way it
is not just used through `verdi daemon` but also through the client.

The locations where the automatic cleaning of stale PID files has also
been changed. Summary of behavior changes:

* `verdi daemon status` and `verdi status` no longer delete stale PID file
* `DaemonClient.start_daemon` is now the only place that deletes them

The `verdi status` and `verdi daemon status` commands were calling the
function `aiida.cmdline.utils.daemon.delete_stale_pid_file` which would
check the daemon PID file and delete it if it was considered stale. This
was added to prevent users having to clean up manually if the daemon got
killed abruptly, leaving behind an orphaned PID file. However, this is a
side-effect that users wouldn't expect from a command that should simply
report the status of a service.

The `aiida.cmdline.utils.daemon.delete_stale_pid_file` function has been
removed and its functionality has been refactored and moved to the
`DaemonClient._check_pid_file` method. It raises if the PID file is
likely to be stale. This allows it to be wrapped by `_is_pid_file_stale`
which returns a boolean, and `_clean_potentially_stale_pid_file` which
replaces the old functionality of automatically deleting the PID file if
it is stale and logs a warning of having done so.

The `daemon.timeout` option default is changed from 20 to 5 seconds.
There is no reason why a functioning daemon wouldn't be able to respond
within 5 seconds, and forcing a user to wait 20 seconds in case of a
broken daemon is frustratingly long.