`verdi process list`: Simplify the daemon load implementation
The calculation of the daemon load at the end of the command was using a
utility function that relied manually parsed the response of the daemon
client. However, the `DaemonClient.get_numprocesses` now does this auto-
matically and `aiida.cmdline.commands.cmd_daemon.execute_client_command`
converts this into the correct formatting for stdout.

The `CalculationQueryBuilder` is replaced with a query using the
`QueryBuilder` directly as we only need the count and with the former we
are forced to define a projection and retrieve the results from the
database to count them in Python, which is more expensive an operation.