CLI: Fail command if `--config` file contains unknown key
Up till now, the `--config` option would accept configuration files with
keys that did not correspond to existing options on the command. This
would lead to incorrect options to be swallowed silently, which can be
surprising to users if they accidentally misspelled an option or used
one that didn't exist.

Here the callback is updated to check the specified keys in the config
and cross-reference them to the params that are defined on the command.
If any unknown keys are specified a `BadParameter` exception is raised
which displays which keys are not supported.

The implementation was originally provided by the `click-config-file`
dependency, however, in order to customize it with the desired behavior
we had to essentially vendor the entire code, since there was no
available hook for the customizations. The code is taken from
https://github.com/phha/click_config_file/blob/7b93a20b4c79458987fac116418859f30a16d82a
which corresponds to `v0.6.0`. Besides the main changes, the default
provider is changed to `yaml_config_file_provider`, the docstrings are
reformated and type hints are added.