default_memory_per_machine_validator should allow 0
The `orm.computer` class has a validator classmethod to check the value
set for "default_memory_per_machine" which is ultmately only used in the
scheduler batch scripts. Currently the validator will only permit
positive integers, which is intuitive, but SLURM allows a value of 0,
interpreting this as "use all memory available". When a scheduler has
this option, `verdi computer setup` prompts the user to use a value, and
so in the case where different queues have differing amounts of memory per
machine, it is desirable to set this default to 0.

This commit changes the condition for the validator from any positive
int to any non-negative int, allowing for this param to be set to 0.