`Scheduler`: Move setting of env variables into base class (#5948)
The `Scheduler` base class implements the concrete method
`_get_submit_script_environment_variables` which formats the lines for
the submission script that set the environment variables that were
defined in the `metadata.options.environment_variables` input.

However, it was left up to the plugins to actually call this method in
the `_get_submit_script_header` and append the lines to the created
submission script header. This would leave it open to plugins forgetting
to add this call and essentially have the environment variables
specified by the caller of the `CalcJob` be ignored.

Here we move the call of `_get_submit_script_environment_variables` to
the `get_submit_script` of the base class, guaranteeing that it will
always be called. This still leaves the possibility to the plugin to
override the method and customize the behavior.

A check is added just before adding the new call, to check whether the
environment variables have already been added indicating the plugin is
also still calling the method. In this case a deprecation warning is
emitted suggesting that the call can be removed from the plugin.

With this change we do assume that setting the environment variables are
idempotent as it is possible that the lines will be included twice until
all external plugins have been updated, but this should be a relatively
safe assumption.