ORM: Add unary operations `+`, `-` and `abs` to `NumericType` (#5946)
Three unary operations are added to the `NumericType` data subclass:

* `+`: Implemented through `__pos__`
* `-`: Implemented through `__neg__`
* `abs`: Implemented through `__abs__`

The `__pos__` dunder overloads the unary plus operator `+` and returns
the value of the object as-is, whereas the `__neg__` overloads the unary
minus operator `-` and returns a new instance of the same class with the
negation of the value of the original object. The `__abs__` dunder is
called when the object is passed to the `abs` built-in and returns the
absolute value.