Fix bug in `aiida.engine.utils.instantiate_process`
If the `process` argument receives an unsupported type that is not a
class, the final check `issubclass(process, Process)` would raise:

    TypeError: issubclass() arg 1 must be a class

which is not very useful to the user. This is fixed by first checking
that the argument is a class to begin with. At least now when an invalid
type is provided, the exception contains the invalid type:

    invalid process <class \'bool\'>, needs to be Process or ProcessBuilder'