`Scheduler`: Refactor interface to make it more generic
The original `Scheduler` interface made the assumption that all
interfaces would interact with the scheduler through a command line
interface that would be invoked through a bash shell. However, this is
not always the case. Prime example is the new FirecREST service, being
developed by CSCS, that will allow to interact with the scheduler
through a REST API. Due to the assumptions of the `Scheduler` interface,
it was difficult to implement it for this use case.

The `Scheduler` interface is made more generic, by removing the
following (abstract) methods:

 * `_get_joblist_command`
 * `_parse_joblist_output`
 * `_get_submit_command`
 * `_parse_submit_output`
 * `submit_from_script`
 * `kill`
 * `_get_kill_command`
 * `_parse_kill_output`

They are replaced by three abstract methods:

 * `submit_job`
 * `get_jobs`
 * `kill_job`

The new interface no longer makes an assumption about how a plugin
implements these methods. The first one should simply submit the job,
given the location of the submission script on the remote computer. The
second should return the status of the list of active jobs. And the
final should kill a job and return the result.

Unfortunately, this change is backwards incompatible and will break
existing scheduler plugins. To simplify the migration pathway, a
subclass `BashCliScheduler` is added. This implements the new `Scheduler`
interface while maintaining the old interface. This means that this new
class is a drop-in replacement of the old `Scheduler` class for existing
plugins. The plugins that ship with `aiida-core` are all updated to
subclass from `BashCliScheduler`. Any existing plugins that subclassed
from these plugins will therefore not be affected whatsoever by these
changes.