Fix the `%verdi` IPython magics utility
The `aiida.tools.ipython.ipython_magics.AiiDALoaderMagics` class
provides the `verdi` method, which allows to invoke `verdi` commands
from a Jupyter notebook, e.g., `%verdi process list`.

This was failing because it would invoke the `verdi` command directly
from the Python API. This would cause the `--profile` parameter to be
evaulated, which would attempt to load the current default profile. If
run from a notebook where the loaded profile was a temporary profile
that had been created on the fly, as is done in the tutorial notebooks
in the documentation, the command would fail. The reason is that the
temporary profile was created and loaded in memory, but not actually
added to the config. The `--profile` parameter, however, would
determine the default profile from the config, which would not be the
loaded temporary profile, and so it would attempt to load a different
profile, which is not allowed, with the default `allow_switch` being set
to `False`.

The solution is to not call the `verdi` command, but the subcommand,
thereby circumventing the `--profile` option. The subcommand name is
retrieved from the parsed command line arguments defined by the `line`
argument, where it should be the first element. A check is performed to
see if the user tried to specify a specific profile, in which case the
first element would be `-p` or `--profile`. This is not supported and an
exception is raised. If any other string is provided that is not a valid
subcommand, the `context.get_command()` call will raise a `UsageError`
just as if it would have been called directly on the CLI.