CI: Fix the benchmark workflow
It was using the `ubuntu-18.04` image which has been decommissioned on
Github actions. It is replaced with `ubuntu-22.04`, pinned and not using
`latest` in order to guarantee reproducibility.

A number of other packages and services are updated:

* Python -> 3.10
* PostgreSQL -> `12.14`
* RabbitMQ -> `3.8.14-management`
* `actions/checkout` -> `v3`
* `actions/setup-python` -> `v4`

A link to the benchmark results is added to the table in the `README.md`.
It points to the new link which will be empty for a while, but will
start to fill up over time.