Dependencies: Bump `graphviz` version to `0.19`
In the commit 80045ae7308de673a838c152998a6257c9a79134 a feature was
added that allows the user to specify the output name of the generated
graph with `verdi node graph generate`. This filename is passed to the
`outfile` input argument of the `graphviz.render()` method, but this was
only added in v0.19 of `graphviz`, see:

https://graphviz.readthedocs.io/en/latest/changelog.html#version-0-19

Here we bump the `graphviz` version to `0.19`.