`DaemonClient`: Fix and homogenize use of `timeout` in client calls (#5960)
The `DaemonClient` provides a number of methods that will attempt to
communicate with the daemon process, e.g., `stop_daemon` and `get_status`.
This is done through the `CircusClient`, as the main daemon process is
the circus daemonizer, which takes a timeout as argument indicating the
number of seconds after which the call should raise if the daemon did
not respond in time.

The default timeout is set in the constructor of the `DaemonClient`
based on the `daemon.timeout` config option. It was incorrectly getting
the global value instead of the profile specific one, which is corrected
by using `config.get_option('daemon.timeout', scope=profile.name)` to
fetch the conifg option.

A `timeout` argument is added to all `DaemonClient` methods that call
through to the `CircusClient` that can be used to override the default
that is based on the `daemon.timeout` config option.

A default is added for `wait` in the `restart_daemon` method, to
homogenize its interface with `start_daemon` and `stop_daemon`. The
manual timeout cycle in `stop_daemon` by calling `_await_condition` has
been removed as this functionality is already performed by the circus
client itself and so is superfluous. The only place it is still used is
in `start_daemon` because there the circus client is not used, since the
circus process is not actually running yet, and a "manual" health check
needs to be performed after the daemon process is launched. The manual
check is added to the `stop_daemon` fixture since the check in certain
unit test scenarios could give a false positive without an additional
manual grace period for `is_daemon_running` to start returning `False`.

The default for the `daemon.timeout` configuration option is decreased
to 2 seconds, as this should be sufficient for most conditions for the
circus daemon process to respond. Note that this daemon process is only
charged with monitoring the daemon workers and so won't be under heavy
load that will prevent it from responding in time.