CLI: Add `--timeout` option to all `verdi daemon` commands
The timeout is passed on to the `DaemonClient` and can be used to
override the default that is defined through the `daemon.timeout`
config option. This can be useful, for example, when the default timeout
set in the configuration is low such that commands don't get stuck
unnecessarily long. However, for certain commands, such as `verdi daemon
stop`, it might make sense to give the client a bit more time to
respond.