`CalcJob`: Fix bug in `validate_calc_job` if `code` excluded
If the `CalcJob` spec is exposed into another process namespace while
excluding the `code` port, the `validate_calc_job` validator defined on
the top-level namespaces would except because it would always attempt to
get the `Computer` through the `computer` attribute from the `code`
input, even if it could be `None`, in the case that the port was
excluded when exposing.

The validator now first checks with `ctx.get_port('code')` whether the
namespace still contains the `code` port and otherwise sets it and
therefore the associated computer to `None`.