`Node`: Validate nested attribute and extra keys
Currently only the top-level keys of the attributes and extras of a node
are validated upon instantiation of the node. Nested keys are only
(silently) converted to a string when storing the node. This behaviour
of silently mutating the keys is dangerous.

Here we add a recursive validation step to the `_validate()` method of
the `Node` class, that relies on the already defined
`validate_attribute_extra_key` function.

Additionally, typing is added to all methods of the `Dict` class.